calcs

+++ include/shfcc/phase_noise_mixer_cc.h
@@ -26,9 +26,9 @@ namespace gr {
        * class. shfcc::phase_noise_mixer_cc::make is the public interface for
        * creating new instances.
        */
-      static sptr make(float fc, float fs, float k0, float k2, float cfo_ampl, float cfo_freq, bool impair);
-      virtual void set_fc(float fc) = 0;
-      virtual void set_fc_nominal(float fc_nom) = 0;
+      static sptr make(double fc, double fs, float k0, float k2, float cfo_ampl, float cfo_freq, bool impair);
+      virtual void set_fc(double fc) = 0;
+      virtual void set_fc_nominal(double fc_nom) = 0;
       virtual void set_k0(float k0) = 0;
       virtual void set_k2(float k2) = 0;
       virtual void set_cfo_ampl(float cfo_ampl) = 0;
